Rename `serializer.py` to `serializers.py`
See also: #37
